refactor(client): refactor rendering into separate file
